use local installation
